Avance de css dimensión tablet, home y primeros pasos muro
