--- src/componentes/Wall.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import { useNavigate } from 'react-router-dom';
-import { logOut } from '../firebase';
-
-function Wall() {
-    const navigate = useNavigate();
-    const handlesSignOut = async () => {
-        await logOut();
-        navigate('/')
-    };
-    return <div>
-       <h2>wall</h2> 
-    <button onClick={handlesSignOut}>Log Out</button>
-    </div>
-}
-
-export default Wall;
