+++ src/style/Home.module.css
@@ -5,10 +5,34 @@ body{
     align-items: center;
 }
 .title{
-    color:antiquewhite;
+    color:#FFB6D5;
     font-size: 60px;
     font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    margin-top: 80px;
 }
 .carroviaje{
     width: 100%;
+
+}
+.login{
+    display:flex;
+    flex-direction: row;
+    align-items: center;
+    margin-bottom: 27px;
+}
+.startText{
+    color:#FFB6D5;
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    font-size: 30px;
+}
+.btnGoggle{
+    width: 240px;
+    height: 60px;
+    margin:20px;
+    border-radius: 15px;
+    background-color: #FFB6D5;
+    border:none;
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    font-size: 25px;
+    font-weight: bold;
 }
