+++ src/style/Wall.module.css
@@ -1,0 +1,17 @@
+body{
+    background-color: #4ca5a4;
+}
+.carHead{
+    width: 100%;
+}
+.btnLogOut{
+    width: 240px;
+    height: 60px;
+    margin:20px;
+    border-radius: 15px;
+    background-color: #FFB6D5;
+    border:none;
+    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    font-size: 23px;
+    font-weight: bold;
+}
